requirements.txt
pandas
sqlalchemy
python-dotenv
prophet
plotly
streamlit
tqdm